Add logic for undoing doInstallSDK
